implemented new uniedit_nameFilter
