+++ utils/suggest_xml_renaming.sh
@@ -3,16 +3,13 @@
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 DST=$(realpath --relative-to="$PWD" "${SCRIPT_DIR}/../dat")
 
-#NAMER=a.out
-NAMER=xml_name.sed
-
 tmp=$(mktemp)
 trap 'rm -f $tmp' EXIT
 
 grep -H -m 1 '^<[a-zA-Z]* name="' $(find $DST -name "*.xml") | tr '[:upper:]' '[:lower:]' |
 sed 's/^\(.*\)\/\([^/]*\)\.xml:.*name=\"\([^"]*\)\".*$/\1\/ \2 \3/' |
 # from now on, we have lines: <path> <filename> <name>
-tee $tmp | cut "-d " -f3- | $SCRIPT_DIR/$NAMER |
+tee $tmp | cut "-d " -f3- | $SCRIPT_DIR/xml_name.sed |
 # rebuild lines with modified names
 paste - $tmp '-d ' | cut '-d ' -f-3 | sed 's/^\([^ ]*\)\ \([^ ]*\)\ \([^ ]*\)/\2 \3 \1/' |
 # keep only lines where <filename> != <name>
